[wip] sqlite as OOP data store
Very much wip, just sharing progress.

Don't forget to run yarn after switching to this branch (requires `better-sqlite3`). (I tested with adding the lib to the site, if this doesnt work try that :)

Run your favorite build commmand with `GATSBY_DB_NODES=sqlite` to switch to this.

Warning: it is slow. But it works.

Consider yourself warned.